:sparkles: feat: user secure activation via email
I thought in a better user activationmethod instead user being stored directly: i let the admin store user with null password, and one hash path is stored for this new user on Cache (Redis) for 24 hours and the user guest receive his hash on email. If the key-value expires in cache memory, the user as a guest can provide a new hash to receive in his email again. With hash in hands, the user guest set his own user password, and after this, he can access the API freely